+++ lib/eneural_net.dart
@@ -1,9 +1,10 @@
+/// eNeural.net library.
 library eneural_net;
 
-export 'src/eneural_net_ann.dart';
 export 'src/eneural_net_activation_functions.dart';
-export 'src/eneural_net_base.dart';
+export 'src/eneural_net_ann.dart';
 export 'src/eneural_net_sample.dart';
 export 'src/eneural_net_scale.dart';
 export 'src/eneural_net_signal.dart';
+export 'src/eneural_net_tools.dart';
 export 'src/eneural_net_training.dart';