+++ lib/eneural_net_extensions.dart
@@ -1,0 +1,5 @@
+/// A group of Dart extensions for numbers and collections
+/// that can help with data handling and calculation.
+library eneural_net_extensions;
+
+export 'src/eneural_net_extension.dart';