+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 1.0.0
 
-- Initial version, created by Stagehand
+- Initial version.
+- Training algorithms: Backpropagation.
+- Activation functions: Sigmoid and approximation versions.
+- Fast math functions.
+- SIMD: Float32x4